Reapeat_type
