+++ lib/utils/app_utils.dart
@@ -1,0 +1,20 @@
+
+import 'dart:ui';
+import 'package:flutter/material.dart';
+
+// All constant in project.
+
+final List<String> categories = ['General', 'Work', 'Health', 'Other'];
+
+final List<int> selectedDays = [];
+
+
+final List<DateTime> selectedDates = [];
+
+final List<Color> predefinedColors = [
+  Colors.orange,
+  Colors.green,
+  Colors.blue,
+  Colors.purple,
+  Colors.red,
+];
