+++ lib/utils/colors.dart
@@ -1,0 +1,31 @@
+
+import 'dart:ui';
+
+class ColorStrings{
+  static const lightSkyBlue =  Color(0xFFD5EEFF) ;
+  static const headingBlue =  Color(0xFF0A3C8E);
+  static const havelockBlue =  Color(0xFFF1F7FD);
+  // static const containerLabelGrey =  Color(9B979817);
+  static const Color whiteColor = Color(0xFFFFFFFF);
+  static const Color blackColor = Color(0xFF000000);
+  // static const Color labelBlackColor = Color(0xFF040415);
+  static const Color redColor = Color(0xFFFF0000);
+  static const Color greenColor = Color(0xFF00FF00);
+  static const Color blueColor = Color(0xFF0000FF);
+  static const Color yellowColor = Color(0xFFFFFF00);
+  static const Color orangeColor = Color(0xFFFFA500);
+  static const Color purpleColor = Color(0xFF800080);
+  static const Color pinkColor = Color(0xFFFFC0CB);
+  static const Color cyanColor = Color(0xFF00FFFF);
+  static const Color magentaColor = Color(0xFFFF00FF);
+  static const Color limeColor = Color(0xFF00FF00);
+  static const Color tealColor = Color(0xFF008080);
+  static const Color indigoColor = Color(0xFF4B0082);
+  static const Color maroonColor = Color(0xFF800000);
+  static const Color turquoiseColor = Color(0xFF40E0D0);
+  static const Color goldColor = Color(0xFFFFD700);
+  static const Color silverColor = Color(0xFFC0C0C0);
+  static const Color grayColor = Color(0xFF808080);
+  static const Color brownColor = Color(0xFFA52A2A);
+
+}
