+++ lib/utils/image_resource.dart
@@ -1,0 +1,10 @@
+
+class ImageResource{
+
+  //String for MainScreen
+  static const String dumbleIcon = 'assets/images/dumble.png';
+  static const String eventIcon = 'assets/images/event.png';
+  static const String calenderIcon = 'assets/images/calenderIcon.svg';
+  static const String closeIcon = 'assets/images/closeIcon.svg';
+
+}
