+++ pubspec.yaml
@@ -46,6 +46,9 @@ dependencies:
   flex_color_picker: ^3.4.1
   font_awesome_flutter: ^10.7.0
   flutter_iconpicker: ^3.3.3
+  flutter_screenutil: ^5.9.3
+  flutter_svg: ^2.0.10+1
+  google_fonts: ^6.2.1
 
 dev_dependencies:
   flutter_test:
@@ -70,9 +73,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - assets/images/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
